Move labeling extraction to new component
