+++ src/client/src/pages/detail/labeling/labelingContext.tsx
@@ -8,6 +8,7 @@ import {
   useLayoutEffect,
   useState,
 } from "react";
+import { DataExtractionResponse } from "../../../api/file/fileInterfaces.ts";
 import {
   ExtractionObject,
   ExtractionState,
@@ -25,6 +26,10 @@ export const LabelingContext = createContext<LabelingContextInterface>({
   setExtractionObject: () => {},
   extractionState: undefined,
   setExtractionState: () => {},
+  fileInfo: undefined,
+  setFileInfo: () => {},
+  setAbortController: () => {},
+  cancelRequest: () => {},
   panelTab: PanelTab.profile,
   setPanelTab: () => {},
 });
@@ -34,11 +39,22 @@ export const LabelingProvider: FC<PropsWithChildren> = ({ children }) => {
   const [panelOpen, setPanelOpen] = useState(false);
   const [extractionObject, setExtractionObject] = useState<ExtractionObject>();
   const [extractionState, setExtractionState] = useState<ExtractionState>();
+  const [fileInfo, setFileInfo] = useState<DataExtractionResponse>();
+  const [abortController, setAbortController] = useState<AbortController>();
   const [panelTab, setPanelTab] = useState<PanelTab>(PanelTab.profile);
 
   const togglePanel = useCallback((isOpen?: boolean) => {
     setPanelOpen(prevState => (isOpen !== undefined ? isOpen : !prevState));
   }, []);
+
+  const cancelRequest = useCallback(() => {
+    if (abortController) {
+      abortController.abort();
+      setAbortController(undefined);
+    }
+    setExtractionObject({ type: "coordinates" });
+    setExtractionState(ExtractionState.start);
+  }, [abortController]);
 
   const panelPositionStorageName = "labelingPanelPosition";
   useLayoutEffect(() => {
@@ -76,6 +92,10 @@ export const LabelingProvider: FC<PropsWithChildren> = ({ children }) => {
         setExtractionObject,
         extractionState,
         setExtractionState,
+        fileInfo,
+        setFileInfo,
+        setAbortController,
+        cancelRequest,
         panelTab,
         setPanelTab,
       }}>
