+++ src/client/src/pages/detail/labeling/labelingInterfaces.tsx
@@ -1,3 +1,4 @@
+import { DataExtractionResponse } from "../../../api/file/fileInterfaces.ts";
 import { ReferenceSystemKey } from "../form/location/coordinateSegmentInterfaces.ts";
 
 export type ExtractionType = "text" | "number" | "coordinates";
@@ -58,6 +59,10 @@ export interface LabelingContextInterface {
   setExtractionObject: (extractionObject: ExtractionObject | undefined) => void;
   extractionState?: ExtractionState;
   setExtractionState: (extractionState: ExtractionState) => void;
+  fileInfo?: DataExtractionResponse;
+  setFileInfo: (fileInfo: DataExtractionResponse) => void;
+  setAbortController: (abortController: AbortController | undefined) => void;
+  cancelRequest: () => void;
   panelTab: PanelTab;
   setPanelTab: (panelTab: PanelTab) => void;
 }
