+++ src/client/src/pages/detail/labeling/labelingPanel.tsx
@@ -1,32 +1,21 @@
-import { FC, MouseEvent, useCallback, useContext, useEffect, useRef, useState } from "react";
+import { FC, MouseEvent, useCallback, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useParams } from "react-router-dom";
 import { Alert, Box, Stack, ToggleButton, ToggleButtonGroup } from "@mui/material";
 import { styled } from "@mui/system";
 import { PanelBottom, PanelRight } from "lucide-react";
-import { ApiError, BoreholeAttachment } from "../../../api/apiInterfaces.ts";
-import { getPhotosByBoreholeId } from "../../../api/fetchApiV2.ts";
-import {
-  extractCoordinates,
-  extractText,
-  fetchExtractionBoundingBoxes,
-  getDataExtractionFileInfo,
-  getFiles,
-  uploadFile,
-} from "../../../api/file/file.ts";
-import { BoreholeFile, DataExtractionResponse, maxFileSizeKB } from "../../../api/file/fileInterfaces.ts";
+import { BoreholeAttachment } from "../../../api/apiInterfaces.ts";
+import { getPhotosByBoreholeId, uploadPhoto } from "../../../api/fetchApiV2.ts";
+import { getFiles, uploadFile } from "../../../api/file/file.ts";
+import { BoreholeFile, maxFileSizeKB } from "../../../api/file/fileInterfaces.ts";
 import { theme } from "../../../AppTheme.ts";
 import { useAlertManager } from "../../../components/alert/alertManager.tsx";
-import { TextExtractionButton } from "../../../components/buttons/labelingButtons.tsx";
-import { DetailContext } from "../detailContext.tsx";
 import { FloatingExtractionFeedback } from "./floatingExtractionFeedback.tsx";
 import { useLabelingContext } from "./labelingContext.tsx";
-import { LabelingDrawContainer } from "./labelingDrawContainer.tsx";
+import { LabelingExtraction } from "./labelingExtraction.tsx";
 import LabelingFileSelector from "./labelingFileSelector.tsx";
 import { LabelingHeader } from "./labelingHeader.tsx";
 import {
-  ExtractionBoundingBox,
-  ExtractionRequest,
   ExtractionState,
   labelingFileFormat,
   matchesFileFormat,
@@ -34,6 +23,7 @@ import {
   PanelTab,
 } from "./labelingInterfaces.tsx";
 import { labelingButtonStyles } from "./labelingStyles.ts";
+import { LabelingView } from "./labelingView.tsx";
 import { PageSelection } from "./pageSelection.tsx";
 
 export const LabelingAlert = styled(Alert)({
@@ -66,27 +56,13 @@ export const LabelingAlert = styled(Alert)({
 const LabelingPanel: FC = () => {
   const { t } = useTranslation();
   const { id: boreholeId } = useParams<{ id: string }>();
-  const {
-    panelPosition,
-    setPanelPosition,
-    extractionObject,
-    setExtractionObject,
-    setExtractionState,
-    extractionState,
-    panelTab,
-  } = useLabelingContext();
+  const { panelPosition, setPanelPosition, extractionState, fileInfo, cancelRequest, panelTab } = useLabelingContext();
   const [isLoadingFiles, setIsLoadingFiles] = useState(true);
   const [files, setFiles] = useState<BoreholeAttachment[]>();
   const [selectedFile, setSelectedFile] = useState<BoreholeAttachment>();
-  const [fileInfo, setFileInfo] = useState<DataExtractionResponse>();
-  const [pageBoundingBoxes, setPageBoundingBoxes] = useState<ExtractionBoundingBox[]>([]);
   const [activePage, setActivePage] = useState<number>(1);
-  const [drawTooltipLabel, setDrawTooltipLabel] = useState<string>();
-  const [extractionExtent, setExtractionExtent] = useState<number[]>([]);
-  const [abortController, setAbortController] = useState<AbortController>();
   const fileInputRef = useRef<HTMLInputElement>(null);
   const { alertIsOpen, text, severity, autoHideDuration, showAlert, closeAlert } = useAlertManager();
-  const { editingEnabled } = useContext(DetailContext);
   const expectedFileFormat = labelingFileFormat[panelTab];
 
   useEffect(() => {
@@ -125,147 +101,30 @@ const LabelingPanel: FC = () => {
 
   const addFile = useCallback(
     async (file: File) => {
-      uploadFile<BoreholeFile>(Number(boreholeId), file)
-        .then(fileResponse => {
-          setSelectedFile(fileResponse.file);
-          loadFiles();
-        })
-        .catch(error => {
-          showAlert(t(error.message), "error");
-        });
-    },
-    [boreholeId, loadFiles, showAlert, t],
-  );
-
-  const setTextToClipboard = useCallback(
-    async (text: string) => {
       try {
-        await navigator.clipboard.writeText(text);
-        const successText = `${t("copiedToClipboard")}: "${text}"`;
-        showAlert(successText.length < 50 ? successText : successText.substring(0, 50) + "...", "info");
-        // eslint-disable-next-line @typescript-eslint/no-unused-vars
-      } catch (err) {
-        showAlert(t("errorCopyingToClipboard"), "error");
+        if (panelTab === PanelTab.profile) {
+          const fileResponse = await uploadFile<BoreholeFile>(Number(boreholeId), file);
+          setSelectedFile(fileResponse.file);
+        } else {
+          const photoResponse = await uploadPhoto(Number(boreholeId), file);
+          setSelectedFile(photoResponse);
+        }
+        loadFiles();
+      } catch (error) {
+        showAlert(t((error as Error).message), "error");
       }
     },
-    [showAlert, t],
+    [boreholeId, loadFiles, panelTab, showAlert, t],
   );
-
-  const triggerDataExtraction = useCallback(
-    (extent: number[]) => {
-      if (fileInfo && extractionObject?.type) {
-        const bbox = {
-          x0: Math.min(...[extent[0], extent[2]]),
-          y0: Math.min(...[extent[1], extent[3]]),
-          x1: Math.max(...[extent[0], extent[2]]),
-          y1: Math.max(...[extent[1], extent[3]]),
-        };
-        setExtractionExtent([]);
-        const request: ExtractionRequest = {
-          filename: fileInfo.fileName.substring(0, fileInfo.fileName.lastIndexOf("-")) + ".pdf",
-          page_number: activePage,
-          bbox: bbox,
-          format: extractionObject.type,
-        };
-        setExtractionState(ExtractionState.loading);
-        setDrawTooltipLabel(undefined);
-        const abortController = new AbortController();
-        setAbortController(abortController);
-        const extractFunction = extractionObject.type === "coordinates" ? extractCoordinates : extractText;
-        extractFunction(request, abortController.signal)
-          .then(response => {
-            if (extractionObject.type) {
-              setExtractionState(ExtractionState.success);
-              setExtractionObject({
-                ...extractionObject,
-                value: response[extractionObject.type],
-              });
-            }
-            if (extractionObject.type === "text") {
-              setTextToClipboard(response[extractionObject.type].toString());
-            }
-          })
-          .catch(error => {
-            if (!error?.toString().includes("AbortError")) {
-              setExtractionState(ExtractionState.error);
-              showAlert(t(error.message), "error");
-            }
-          })
-          .finally(() => {
-            setAbortController(undefined);
-          });
-      }
-    },
-    [activePage, extractionObject, fileInfo, setExtractionObject, setExtractionState, setTextToClipboard, showAlert, t],
-  );
-
-  useEffect(() => {
-    if (extractionExtent?.length > 0) {
-      triggerDataExtraction(extractionExtent);
-    }
-  }, [extractionExtent, triggerDataExtraction]);
-
-  const cancelRequest = () => {
-    if (abortController) {
-      abortController.abort();
-      setAbortController(undefined);
-    }
-    setExtractionObject({ type: "coordinates" });
-    setExtractionState(ExtractionState.start);
-    setExtractionExtent([]);
-  };
 
   useEffect(() => {
     loadFiles();
   }, [loadFiles]);
 
   useEffect(() => {
-    if (extractionState === ExtractionState.start) {
-      closeAlert();
-      setExtractionState(ExtractionState.drawing);
-      if (extractionObject?.type === "coordinates") {
-        setDrawTooltipLabel("drawCoordinateBox");
-      }
-    }
-  }, [closeAlert, extractionObject, extractionState, setExtractionObject, setExtractionState]);
-
-  useEffect(() => {
-    if (!selectedFile) return;
-
-    const fetchExtractionData = async () => {
-      const fileInfoResponse = await getDataExtractionFileInfo(selectedFile.id, activePage);
-      const { fileName, count } = fileInfoResponse;
-      let newActivePage = activePage;
-      if (fileInfo?.count !== count) {
-        newActivePage = 1;
-        setActivePage(newActivePage);
-      }
-      if (fileInfo?.fileName !== fileName) {
-        setFileInfo(fileInfoResponse);
-        try {
-          const boundingBoxResponse = await fetchExtractionBoundingBoxes(selectedFile.nameUuid, newActivePage);
-          setPageBoundingBoxes(boundingBoxResponse.bounding_boxes);
-        } catch (error) {
-          if (error instanceof ApiError) {
-            showAlert(t(error.message), "warning");
-          } else {
-            showAlert(t("errorDataExtractionFetchBoundingBoxes"), "warning");
-          }
-        }
-      }
-    };
-
-    fetchExtractionData();
-  }, [activePage, selectedFile, fileInfo?.count, fileInfo?.fileName, showAlert, t, editingEnabled]);
-
-  useEffect(() => {
-    if (files && files.length > 0 && (panelTab === PanelTab.photo || files.length === 1)) {
-      setSelectedFile(files[0]);
-    } else {
-      setSelectedFile(undefined);
-    }
+    setSelectedFile(undefined);
     setActivePage(1);
-  }, [files, panelTab]);
+  }, [panelTab]);
 
   const isExtractionLoading = extractionState === ExtractionState.loading;
   return (
@@ -351,24 +210,6 @@ const LabelingPanel: FC = () => {
       </Stack>
       {selectedFile ? (
         <Box sx={{ height: "100%", width: "100%", position: "relative" }}>
-          <Box
-            sx={{
-              position: "absolute",
-              top: theme.spacing(2),
-              left: theme.spacing(2),
-              zIndex: "500",
-            }}>
-            {editingEnabled && (
-              <TextExtractionButton
-                disabled={extractionObject?.type == "text" && extractionState === ExtractionState.drawing}
-                onClick={() => {
-                  setExtractionObject({ type: "text" });
-                  setExtractionState(ExtractionState.start);
-                  setDrawTooltipLabel("drawTextBox");
-                }}
-              />
-            )}
-          </Box>
           <FloatingExtractionFeedback
             isExtractionLoading={isExtractionLoading}
             cancelRequest={cancelRequest}
@@ -377,13 +218,17 @@ const LabelingPanel: FC = () => {
             closeAlert={closeAlert}
             alertIsOpen={alertIsOpen}
           />
-          <LabelingDrawContainer
-            fileInfo={fileInfo}
-            onDrawEnd={setExtractionExtent}
-            drawTooltipLabel={drawTooltipLabel}
-            boundingBoxes={pageBoundingBoxes}
-            extractionType={extractionObject?.type}
-          />
+          {panelTab === PanelTab.profile ? (
+            <LabelingExtraction
+              selectedFile={selectedFile}
+              activePage={activePage}
+              setActivePage={setActivePage}
+              showAlert={showAlert}
+              closeAlert={closeAlert}
+            />
+          ) : (
+            <LabelingView fileInfo={undefined} loadImage={async () => null} />
+          )}
         </Box>
       ) : (
         <LabelingFileSelector
